refactor: remove useBadge to its own folder.
