+++ packages/mui-base/src/BadgeUnstyled/index.ts
@@ -1,8 +1,5 @@
 export { default } from './BadgeUnstyled';
 
-export { default as useBadge } from './useBadge';
-export * from './useBadge';
-
 export * from './BadgeUnstyled.types';
 
 export { default as badgeUnstyledClasses } from './badgeUnstyledClasses';
