+++ packages/mui-base/src/useBadge/index.ts
@@ -1,0 +1,2 @@
+export { default as useBadge } from './useBadge';
+export * from './useBadge';
